+++ src/bpp/tests/test_models/test_abstract.py
@@ -11,14 +11,20 @@ from bpp.models.seria_wydawnicza import Seria_Wydawnicza
 from bpp.models.struktura import Jednostka, Wydzial
 from bpp.models.system import Charakter_PBN, Charakter_Formalny, Typ_KBN
 from bpp.models.wydawnictwo_ciagle import Wydawnictwo_Ciagle
-from bpp.models.wydawnictwo_zwarte import Wydawnictwo_Zwarte_Autor, \
-    Wydawnictwo_Zwarte
+from bpp.models.wydawnictwo_zwarte import Wydawnictwo_Zwarte_Autor, Wydawnictwo_Zwarte
 
 
 @pytest.mark.django_db
 def test_baza_modelu_odpowiedzialnosci_zapisywanie(
-        wydawnictwo_ciagle, autor_jan_nowak, rok, jednostka,
-        dyscyplina1, dyscyplina2, typy_odpowiedzialnosci, db):
+    wydawnictwo_ciagle,
+    autor_jan_nowak,
+    rok,
+    jednostka,
+    dyscyplina1,
+    dyscyplina2,
+    typy_odpowiedzialnosci,
+    db,
+):
     wydawnictwo_ciagle.rok = rok
     wydawnictwo_ciagle.save()
 
@@ -35,7 +41,7 @@ def test_baza_modelu_odpowiedzialnosci_zapisywanie(
         jednostka=jednostka,
         typ_odpowiedzialnosci=Typ_Odpowiedzialnosci.objects.get(skrot="aut."),
         zapisany_jako="Foobar",
-        dyscyplina_naukowa=None
+        dyscyplina_naukowa=None,
     )
     wca.save()
 
@@ -63,22 +69,28 @@ def test_baza_modelu_odpowiedzialnosci_zapisywanie(
 
 @pytest.mark.django_db
 def test_baza_modelu_odpowiedzialnosci_autorow_dyscyplina_okresl_dyscypline(
-        wydawnictwo_ciagle,
-        jednostka,
-        autor_jan_kowalski,
-        dyscyplina1,
-        dyscyplina2,
-        typy_odpowiedzialnosci, rok):
+    wydawnictwo_ciagle,
+    jednostka,
+    autor_jan_kowalski,
+    dyscyplina1,
+    dyscyplina2,
+    typy_odpowiedzialnosci,
+    rok,
+):
 
-    wca = wydawnictwo_ciagle.dodaj_autora(autor_jan_kowalski, jednostka, zapisany_jako="Kowalski")
-    assert wca.okresl_dyscypline() == None
+    wca = wydawnictwo_ciagle.dodaj_autora(
+        autor_jan_kowalski, jednostka, zapisany_jako="Kowalski"
+    )
+    assert wca.okresl_dyscypline() is None
 
     Autor_Dyscyplina.objects.create(
-        rok=rok, autor=autor_jan_kowalski,
-        dyscyplina_naukowa=dyscyplina1
+        rok=rok,
+        autor=autor_jan_kowalski,
+        dyscyplina_naukowa=dyscyplina1,
+        subdyscyplina_naukowa=dyscyplina2,
     )
 
-    assert wca.okresl_dyscypline() == None
+    assert wca.okresl_dyscypline() is None
 
     wca.dyscyplina_naukowa = dyscyplina2
     wca.save()
@@ -86,42 +98,37 @@ def test_baza_modelu_odpowiedzialnosci_autorow_dyscyplina_okresl_dyscypline(
 
     wca.dyscyplina_naukowa = None
     wca.save()
-    assert wca.okresl_dyscypline() == None
+    assert wca.okresl_dyscypline() is None
 
 
 @pytest.mark.django_db
 def test_eksport_pbn_outstanding():
-    wz = mommy.make(Wydawnictwo_Zwarte, praca_wybitna=True,
-                    uzasadnienie_wybitnosci="foobar")
-    toplevel = Element('test')
+    wz = mommy.make(
+        Wydawnictwo_Zwarte, praca_wybitna=True, uzasadnienie_wybitnosci="foobar"
+    )
+    toplevel = Element("test")
     wz.eksport_pbn_outstanding(toplevel)
     assert len(toplevel.getchildren()) == 2
 
     wz.praca_wybitna = False
-    toplevel = Element('test')
+    toplevel = Element("test")
     wz.eksport_pbn_outstanding(toplevel)
-    assert len(toplevel.getchildren()) == 0
+    assert len(toplevel.getchildren()) == 1
 
 
 @pytest.mark.django_db
-def test_eksport_pbn_outstanding():
+def test_eksport_pbn_outstanding_2():
     wz = mommy.make(Wydawnictwo_Zwarte)
 
-    toplevel = Element('test')
+    toplevel = Element("test")
     wz.eksport_pbn_award(toplevel)
     assert len(toplevel.getchildren()) == 0
 
-    n1 = mommy.make(Nagroda,
-                    object=wz,
-                    rok_przyznania=2000,
-                    uzasadnienie="foobar")
+    mommy.make(Nagroda, object=wz, rok_przyznania=2000, uzasadnienie="foobar")
 
-    n1 = mommy.make(Nagroda,
-                    object=wz,
-                    rok_przyznania=2001,
-                    uzasadnienie="baz quux")
+    mommy.make(Nagroda, object=wz, rok_przyznania=2001, uzasadnienie="baz quux")
 
-    toplevel = Element('test')
+    toplevel = Element("test")
     wz.eksport_pbn_award(toplevel)
     assert len(toplevel.getchildren()) == 2
 
@@ -130,12 +137,12 @@ def test_eksport_pbn_outstanding():
 def test_eksport_pbn_issn():
     wz = mommy.make(Wydawnictwo_Zwarte, issn="foobar")
 
-    toplevel = Element('test')
+    toplevel = Element("test")
     wz.eksport_pbn_issn(toplevel)
     assert len(toplevel.getchildren())
 
     wz.issn = None
-    toplevel = Element('test')
+    toplevel = Element("test")
     wz.eksport_pbn_issn(toplevel)
     assert len(toplevel.getchildren()) == 0
 
@@ -145,7 +152,7 @@ def test_eksport_pbn_seria():
     s = mommy.make(Seria_Wydawnicza, nazwa="TEST")
     wz = mommy.make(Wydawnictwo_Zwarte, seria_wydawnicza=s, numer_w_serii="15")
 
-    toplevel = Element('test')
+    toplevel = Element("test")
     wz.eksport_pbn_series(toplevel)
     wz.eksport_pbn_number_in_series(toplevel)
     assert len(toplevel.getchildren()) == 2
@@ -154,7 +161,7 @@ def test_eksport_pbn_seria():
     wz.numer_w_serii = None
     wz.save()
 
-    toplevel = Element('test')
+    toplevel = Element("test")
     wz.eksport_pbn_series(toplevel)
     wz.eksport_pbn_number_in_series(toplevel)
     assert len(toplevel.getchildren()) == 0
@@ -164,38 +171,36 @@ def test_eksport_pbn_seria():
 def test_eksport_pbn_conference():
     konf = mommy.make(Konferencja)
     wc = mommy.make(Wydawnictwo_Zwarte, konferencja=konf)
-    toplevel = Element('test')
+    toplevel = Element("test")
     wc.eksport_pbn_conference(toplevel)
     assert len(toplevel.getchildren())
 
 
 @pytest.mark.django_db
 def test_eksport_pbn_is():
-    c_pbn = mommy.make(Charakter_PBN, identyfikator='fubar')
+    c_pbn = mommy.make(Charakter_PBN, identyfikator="fubar")
     c = mommy.make(Charakter_Formalny, charakter_pbn=c_pbn)
 
-    c_pbn2 = mommy.make(Charakter_PBN, identyfikator='baz')
+    c_pbn2 = mommy.make(Charakter_PBN, identyfikator="baz")
     t = mommy.make(Typ_KBN, charakter_pbn=c_pbn2)
 
-    wz = mommy.make(Wydawnictwo_Zwarte,
-                    charakter_formalny=c,
-                    typ_kbn=t)
+    wz = mommy.make(Wydawnictwo_Zwarte, charakter_formalny=c, typ_kbn=t)
 
-    toplevel = Element('test')
+    toplevel = Element("test")
     wz.eksport_pbn_is(toplevel=toplevel)
     assert toplevel.getchildren()[0].text == "fubar"
 
     c.charakter_pbn = None
     c.save()
 
-    toplevel = Element('test')
+    toplevel = Element("test")
     wz.eksport_pbn_is(toplevel=toplevel)
     assert toplevel.getchildren()[0].text == "baz"
 
     t.charakter_pbn = None
     t.save()
 
-    toplevel = Element('test')
+    toplevel = Element("test")
     wz.eksport_pbn_is(toplevel=toplevel)
     with pytest.raises(IndexError):
         assert toplevel.getchildren()[0]
@@ -203,22 +208,24 @@ def test_eksport_pbn_is():
 
 @pytest.mark.parametrize(
     "klass, pbn_id,expected",
-    [(Wydawnictwo_Ciagle, 500, "500"),
-     (Wydawnictwo_Ciagle, None, "2000000100"),
-     (Wydawnictwo_Zwarte, None, "4000000100")
-     ]
+    [
+        (Wydawnictwo_Ciagle, 500, "500"),
+        (Wydawnictwo_Ciagle, None, "2000000100"),
+        (Wydawnictwo_Zwarte, None, "4000000100"),
+    ],
 )
 @pytest.mark.django_db
 def test_eksport_pbn_system_identifier(klass, pbn_id, expected):
     wc = mommy.make(klass, pbn_id=pbn_id, pk=100)
-    toplevel = Element('test')
+    toplevel = Element("test")
     wc.eksport_pbn_system_identifier(toplevel=toplevel)
     assert toplevel.getchildren()[0].text == expected
 
 
 @pytest.mark.django_db
-def test_eksport_pbn_author_afiliacja_w_kontekscie_wydzialu(uczelnia,
-                                                            autor_jan_kowalski, wydawnictwo_zwarte, standard_data):
+def test_eksport_pbn_author_afiliacja_w_kontekscie_wydzialu(
+    uczelnia, autor_jan_kowalski, wydawnictwo_zwarte, standard_data
+):
     w1 = mommy.make(Wydzial, uczelnia=uczelnia)
     w2 = mommy.make(Wydzial, uczelnia=uczelnia)
 
@@ -245,8 +252,8 @@ def test_eksport_pbn_author_afiliacja_w_kontekscie_wydzialu(uczelnia,
 
 @pytest.mark.django_db
 def test_eksport_pbn_editor_afiliacja_w_kontekscie_wydzialu(
-        uczelnia, autor_jan_kowalski, wydawnictwo_zwarte,
-        typy_odpowiedzialnosci):
+    uczelnia, autor_jan_kowalski, wydawnictwo_zwarte, typy_odpowiedzialnosci
+):
     w1 = mommy.make(Wydzial, uczelnia=uczelnia)
     w2 = mommy.make(Wydzial, uczelnia=uczelnia)
 
@@ -258,7 +265,9 @@ def test_eksport_pbn_editor_afiliacja_w_kontekscie_wydzialu(
     autor_jan_kowalski.dodaj_jednostke(j2)
 
     # W pracy określ afiliację do drugiego wydziału
-    wydawnictwo_zwarte.dodaj_autora(autor_jan_kowalski, j2, typ_odpowiedzialnosci_skrot="red.")
+    wydawnictwo_zwarte.dodaj_autora(
+        autor_jan_kowalski, j2, typ_odpowiedzialnosci_skrot="red."
+    )
 
     # I teraz ma wyrzucić wszystkich redaktorów
     toplevel = []
@@ -269,7 +278,6 @@ def test_eksport_pbn_editor_afiliacja_w_kontekscie_wydzialu(
     toplevel = []
     wydawnictwo_zwarte.eksport_pbn_editor(toplevel, w2, Wydawnictwo_Zwarte_Autor)
     assert len(toplevel) == 1
-
 
 
 @pytest.mark.django_db
@@ -297,8 +305,9 @@ def test_eksport_pbn_editor_afiliacja_w_kontekscie_wydzialu(
         ("s. P-29.", "P-29"),
         ("s. PP-4.", "PP-4"),
         # ("s. P-27-P-28.", "P-27-P-28"), # to się nie uda
-        ("s. xiii-xiv.", "xiii-xiv")
-    ])
+        ("s. xiii-xiv.", "xiii-xiv"),
+    ],
+)
 def test_eksport_pbn_zakres_stron(input, expected, wydawnictwo_ciagle):
     wydawnictwo_ciagle.szczegoly = input
     ret = wydawnictwo_ciagle.eksport_pbn_zakres_stron()